# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FIRS E-Invoicing Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List all invoices
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create a new invoice (draft)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceRequest"
      responses:
        "201":
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceRequest"
      responses:
        "200":
          description: Invoice updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}/submit:
    post:
      operationId: submitInvoice
      tags: [invoices]
      summary: Submit invoice to FIRS
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submission result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionResponse"
  /invoices/{id}/validate:
    post:
      operationId: validateInvoice
      tags: [invoices]
      summary: Validate invoice before submission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    Customer:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        tin:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        address:
          type: string
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - tin
        - address
...
    LineItem:
      type: object
      properties:
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        vatRate:
          type: number
        amount:
          type: number
      required:
        - description
        - quantity
        - unitPrice
        - vatRate
        - amount
    InvoiceRequest:
      type: object
      properties:
        customerId:
          type: integer
        invoiceNumber:
          type: string
        invoiceDate:
          type: string
          format: date
        dueDate:
          type: string
          format: date
          nullable: true
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/LineItem"
        subtotal:
          type: number
        vatAmount:
          type: number
        totalAmount:
          type: number
        currency:
          type: string
          default: NGN
        notes:
          type: string
          nullable: true
      required:
        - customerId
        - invoiceNumber
        - invoiceDate
        - lineItems
        - subtotal
        - vatAmount
        - totalAmount
    Invoice:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        customerId:
          type: integer
        customer:
          $ref: "#/components/schemas/Customer"
          nullable: true
        invoiceNumber:
          type: string
        invoiceDate:
          type: string
        dueDate:
          type: string
          nullable: true
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/LineItem"
        subtotal:
          type: number
        vatAmount:
          type: number
        totalAmount:
          type: number
        currency:
          type: string
        status:
          type: string
          enum: [draft, pending, accepted, rejected]
        firsReferenceId:
          type: string
          nullable: true
        rejectionReasons:
          type: array
          items:
            type: string
          nullable: true
        notes:
          type: string
          nullable: true
        submittedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - customerId
        - invoiceNumber
        - invoiceDate
        - lineItems
        - subtotal
        - vatAmount
        - totalAmount
        - currency
        - status
    ValidationResponse:
      type: object
      properties:
        valid:
          type: boolean
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ValidationError"
      required:
        - valid
        - errors
    ValidationError:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
        code:
          type: string
      required:
        - field
        - message
        - code
    SubmissionResponse:
      type: object
      properties:
        success:
          type: boolean
        status:
          type: string
          enum: [accepted, rejected]
        firsReferenceId:
          type: string
          nullable: true
        message:
          type: string
        rejectionReasons:
          type: array
          items:
            type: string
          nullable: true
        submittedAt:
          type: string
          format: date-time
      required:
        - success
        - status
        - message
        - submittedAt