# lib/api-zod/src/generated/types/invoice.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FIRS E-Invoicing Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Customer } from "./customer";
import type { InvoiceStatus } from "./invoiceStatus";
import type { LineItem } from "./lineItem";

export interface Invoice {
  id: number;
  userId: number;
  customerId: number;
  customer?: Customer | null;
  invoiceNumber: string;
  invoiceDate: string;
  dueDate?: string | null;
  lineItems: LineItem[];
  subtotal: number;
  vatAmount: number;
  totalAmount: number;
  currency: string;
  status: InvoiceStatus;
  firsReferenceId?: string | null;
  rejectionReasons?: string[] | null;
  notes?: string | null;
  submittedAt?: Date | null;
  createdAt?: Date;
  updatedAt?: Date;
}