# lib/api-zod/src/generated/types/submissionResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FIRS E-Invoicing Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { SubmissionResponseStatus } from "./submissionResponseStatus";

export interface SubmissionResponse {
  success: boolean;
  status: SubmissionResponseStatus;
  firsReferenceId?: string | null;
  message: string;
  rejectionReasons?: string[] | null;
  submittedAt: Date;
}